requirements.txt
pydicom
scipy
numpy
onnxruntime
Pillow
